*: move bio* under its own pkg
